Improved insertion and deletion of pods from cluster
Using what is called experience I've figured out how to add new pods
to a running cluster. Given that only the pods that are ready will be
under LB, asking directly LB to add new pod to the cluster makes it
whole thing really easy.

The only problem is to create new cluster when no pod is in one
yet. This is a special case scenario, I'm yet to solve.